Refactor ThinDevId impl into a separate module
For better encapsulation and testing; it should make other refactorings
easier.
